LPS-53879 Disable deletion of temporal LAR using portal-ext.properties

+++ portal-impl/src/com/liferay/portal/lar/backgroundtask/LayoutStagingBackgroundTaskExecutor.java
@@ -38,6 +38,7 @@ import com.liferay.portal.service.LayoutSetBranchLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.StagingLocalServiceUtil;
 import com.liferay.portal.spring.transaction.TransactionHandlerUtil;
+import com.liferay.portal.util.PropsValues;
 
 import java.io.File;
 import java.io.Serializable;
@@ -217,8 +218,25 @@ public class LayoutStagingBackgroundTaskExecutor
 
 				initLayoutSetBranches(_userId, _sourceGroupId, _targetGroupId);
 			}
-			finally {
+			catch (Exception e) {
+				if (PropsValues.STAGING_DELETE_TEMP_LAR_ON_FAILURE) {
+					FileUtil.delete(file);
+				}
+				else if (file != null) {
+					_log.error(
+						"Temporary LAR file is kept at " +
+							file.getAbsolutePath());
+				}
+
+				throw e;
+			}
+
+			if (PropsValues.STAGING_DELETE_TEMP_LAR_ON_SUCCESS) {
 				FileUtil.delete(file);
+			}
+			else if (file != null) {
+				_log.error(
+					"Temporary LAR file is kept at " + file.getAbsolutePath());
 			}
 
 			return missingReferences;
